+++ mmct/video_pipeline/__init__.py
@@ -7,11 +7,9 @@ This package exposes the main public API for video ingestion and querying.
 from .agents.video_agent import VideoAgent
 from .core.ingestion.ingestion_pipeline import IngestionPipeline
 from .core.ingestion.languages import Languages
-from .core.ingestion.transcription.transcription_services import TranscriptionServices
 
 __all__ = [
     "VideoAgent",
     "IngestionPipeline",
     "Languages",
-    "TranscriptionServices",
 ]
