+++ mmct/video_pipeline/core/ingestion/pipelines/steps/chapters/semantic_chunking/__init__.py
@@ -4,7 +4,7 @@ Semantic Chunking Module
 This module provides semantic clustering functionality for video transcripts.
 """
 
-from mmct.video_pipeline.core.ingestion.semantic_chunking.semantic_chunker import SemanticChunker
+from .semantic_chunker import SemanticChunker
 
 __all__ = [
     "SemanticChunker",
