+++ pyproject.toml
@@ -81,7 +81,7 @@ video-agent = [
     "opencv-python-headless>=4.8.0.0,<5.0.0",
     "ffmpeg-python",
     "pydub>=0.25.1,<1.0.0",
-    "pyaudioop>=0.2.2",
+    "audioop-lts>=0.2.1",
     "scikit-learn>=1.3.0,<2.0.0",
     "scipy>=1.11.0,<2.0.0",
     "networkx",
